feat: import ExperienceType and ThemeVariant types in VersionSelectionPage
Clean up unused components by removing:
- dynamic-landing-page.tsx
- landing-page-with-header-footer.tsx
- dynamic-header.tsx
- enhanced-header.tsx
- standard-landing-page-screenshot-match.tsx
- simple-header.tsx

Update import paths for ThemeToggle and fix component references in layout files.
