+++ src/components/layout/StandardMobileHeader.tsx
@@ -1,12 +1,10 @@
 // src/components/layout/StandardMobileHeader.tsx
 'use client'
 
-import React from 'react'
 import { Button } from '@/components/ui/button'
-import { HeaderNavigationItem, NavigationSubItem } from '@/lib/navigation'
-import { cn } from '@/lib/utils'
 import { useAnalytics } from '@/hooks/use-analytics'
-+import { EventCategory } from '@/lib/analytics/constants' 
+import { HeaderNavigationItem } from '@/lib/navigation'
+import { cn } from '@/lib/utils'
 import Link from 'next/link'
 
 export interface StandardMobileHeaderProps {
